Created using spr 1.3.5

[skip ci]

+++ core/internal/runhistory/runhistory_test.go
@@ -31,10 +31,10 @@ func TestApplyUpdate(t *testing.T) {
 		})
 
 	expectedTree := pathtree.TreeData{
-		"setting1": int64(69),
+		"setting1": float64(69),
 		"config": pathtree.TreeData{
 			"setting2": pathtree.TreeData{
-				"value": int64(42),
+				"value": float64(42),
 			},
 		},
 	}
